Add files via upload
This file contains an exe file which reads from config file to postprocess the xyz file from a lammps run. After that it will do a kde mapping to a 2d grid and write the results in a csv file. That csv file can be used to create density plot.